Twinder local server and client working finally.
After plenty of debug. First I narrowed down the problem to the server
itself. Then it was the java class not found, which is solved by go to
intelliJ project structure to put all to output. Then it was the getBytes
method, which doesnt recognize UTF-8 keyword. Then the most tricky part
is the basicPublish method, which should NOT take fanoutExchange keyword
in the first parameter location in this case.