LAB08: made Consumer work
Not a easy task. Turns out that the consumer can not decode the message
because the server is not sending expected format.
This is the reason why we need that additional SwipeBody class... the
key part is the toString method, which format the String using
Gson.toJSON. Key trick.
On a high level, it is kind of obvious, you need to package the message
using GSON on server, so that the consumer can use GSON with same class
to decode.